updates Drupal to 10.1 + contrib modules

+++ web/core/modules/workspaces/src/WorkspaceAssociation.php
@@ -6,11 +6,16 @@ use Drupal\Core\Database\Connection;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Entity\RevisionableInterface;
 use Drupal\Core\Entity\Sql\SqlContentEntityStorage;
+use Drupal\Core\Utility\Error;
+use Drupal\workspaces\Event\WorkspacePostPublishEvent;
+use Drupal\workspaces\Event\WorkspacePublishEvent;
+use Psr\Log\LoggerInterface;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 /**
  * Provides a class for CRUD operations on workspace associations.
  */
-class WorkspaceAssociation implements WorkspaceAssociationInterface {
+class WorkspaceAssociation implements WorkspaceAssociationInterface, EventSubscriberInterface {
 
   /**
    * The table for the workspace association storage.
@@ -47,11 +52,17 @@ class WorkspaceAssociation implements WorkspaceAssociationInterface {
    *   The entity type manager for querying revisions.
    * @param \Drupal\workspaces\WorkspaceRepositoryInterface $workspace_repository
    *   The Workspace repository service.
+   * @param \Psr\Log\LoggerInterface|null $logger
+   *   The logger.
    */
-  public function __construct(Connection $connection, EntityTypeManagerInterface $entity_type_manager, WorkspaceRepositoryInterface $workspace_repository) {
+  public function __construct(Connection $connection, EntityTypeManagerInterface $entity_type_manager, WorkspaceRepositoryInterface $workspace_repository, protected ?LoggerInterface $logger = NULL) {
     $this->database = $connection;
     $this->entityTypeManager = $entity_type_manager;
     $this->workspaceRepository = $workspace_repository;
+    if ($this->logger === NULL) {
+      @trigger_error('Calling ' . __METHOD__ . '() without the $logger argument is deprecated in drupal:10.1.0 and it will be required in drupal:11.0.0. See https://www.drupal.org/node/2932520', E_USER_DEPRECATED);
+      $this->logger = \Drupal::service('logger.channel.workspaces');
+    }
   }
 
   /**
@@ -113,7 +124,7 @@ class WorkspaceAssociation implements WorkspaceAssociationInterface {
       if (isset($transaction)) {
         $transaction->rollBack();
       }
-      watchdog_exception('workspaces', $e);
+      Error::logException($this->logger, $e);
       throw $e;
     }
   }
@@ -175,18 +186,62 @@ class WorkspaceAssociation implements WorkspaceAssociationInterface {
     $id_field = $table_mapping->getColumnNames($entity_type->getKey('id'))['value'];
     $revision_id_field = $table_mapping->getColumnNames($entity_type->getKey('revision'))['value'];
 
+    $workspace_tree = $this->workspaceRepository->loadTree();
+    if (isset($workspace_tree[$workspace_id])) {
+      $workspace_candidates = array_merge([$workspace_id], $workspace_tree[$workspace_id]['ancestors']);
+    }
+    else {
+      $workspace_candidates = [$workspace_id];
+    }
+
     $query = $this->database->select($entity_type->getRevisionTable(), 'revision');
     $query->leftJoin($entity_type->getBaseTable(), 'base', "[revision].[$id_field] = [base].[$id_field]");
 
     $query
       ->fields('revision', [$revision_id_field, $id_field])
-      ->condition("revision.$workspace_field", $workspace_id)
-      ->where("[revision].[$revision_id_field] > [base].[$revision_id_field]")
+      ->condition("revision.$workspace_field", $workspace_candidates, 'IN')
+      ->where("[revision].[$revision_id_field] >= [base].[$revision_id_field]")
       ->orderBy("revision.$revision_id_field", 'ASC');
 
     // Restrict the result to a set of entity ID's if provided.
     if ($entity_ids) {
       $query->condition("revision.$id_field", $entity_ids, 'IN');
+    }
+
+    return $query->execute()->fetchAllKeyed();
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getAssociatedInitialRevisions(string $workspace_id, string $entity_type_id, array $entity_ids = []) {
+    /** @var \Drupal\Core\Entity\EntityStorageInterface $storage */
+    $storage = $this->entityTypeManager->getStorage($entity_type_id);
+
+    // If the entity type is not using core's default entity storage, we can't
+    // assume the table mapping layout so we have to return only the latest
+    // tracked revisions.
+    if (!$storage instanceof SqlContentEntityStorage) {
+      return $this->getTrackedEntities($workspace_id, $entity_type_id, $entity_ids)[$entity_type_id];
+    }
+
+    $entity_type = $storage->getEntityType();
+    $table_mapping = $storage->getTableMapping();
+    $workspace_field = $table_mapping->getColumnNames($entity_type->get('revision_metadata_keys')['workspace'])['target_id'];
+    $id_field = $table_mapping->getColumnNames($entity_type->getKey('id'))['value'];
+    $revision_id_field = $table_mapping->getColumnNames($entity_type->getKey('revision'))['value'];
+
+    $query = $this->database->select($entity_type->getBaseTable(), 'base');
+    $query->leftJoin($entity_type->getRevisionTable(), 'revision', "[base].[$revision_id_field] = [revision].[$revision_id_field]");
+
+    $query
+      ->fields('base', [$revision_id_field, $id_field])
+      ->condition("revision.$workspace_field", $workspace_id, '=')
+      ->orderBy("base.$revision_id_field", 'ASC');
+
+    // Restrict the result to a set of entity ID's if provided.
+    if ($entity_ids) {
+      $query->condition("base.$id_field", $entity_ids, 'IN');
     }
 
     return $query->execute()->fetchAllKeyed();
@@ -208,21 +263,37 @@ class WorkspaceAssociation implements WorkspaceAssociationInterface {
    * {@inheritdoc}
    */
   public function postPublish(WorkspaceInterface $workspace) {
+    @trigger_error(__METHOD__ . '() is deprecated in drupal:10.1.0 and is removed from drupal:11.0.0. Use the \Drupal\workspaces\Event\WorkspacePostPublishEvent event instead. See https://www.drupal.org/node/3242573', E_USER_DEPRECATED);
     $this->deleteAssociations($workspace->id());
   }
 
   /**
    * {@inheritdoc}
    */
-  public function deleteAssociations($workspace_id, $entity_type_id = NULL, $entity_ids = NULL) {
-    $query = $this->database->delete(static::TABLE)
-      ->condition('workspace', $workspace_id);
+  public function deleteAssociations($workspace_id = NULL, $entity_type_id = NULL, $entity_ids = NULL, $revision_ids = NULL) {
+    if (!$workspace_id && !$entity_type_id) {
+      throw new \InvalidArgumentException('A workspace ID or an entity type ID must be provided.');
+    }
+
+    $query = $this->database->delete(static::TABLE);
+
+    if ($workspace_id) {
+      $query->condition('workspace', $workspace_id);
+    }
 
     if ($entity_type_id) {
+      if (!$entity_ids && !$revision_ids) {
+        throw new \InvalidArgumentException('A list of entity IDs or revision IDs must be provided for an entity type.');
+      }
+
       $query->condition('target_entity_type_id', $entity_type_id, '=');
 
       if ($entity_ids) {
         $query->condition('target_entity_id', $entity_ids, 'IN');
+      }
+
+      if ($revision_ids) {
+        $query->condition('target_entity_revision_id', $revision_ids, 'IN');
       }
     }
 
@@ -246,6 +317,25 @@ class WorkspaceAssociation implements WorkspaceAssociationInterface {
       $indexed_rows->condition('workspace', $parent_id);
       $this->database->insert(static::TABLE)->from($indexed_rows)->execute();
     }
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function getSubscribedEvents(): array {
+    // Workspace association records cleanup should happen as late as possible.
+    $events[WorkspacePostPublishEvent::class][] = ['onPostPublish', -500];
+    return $events;
+  }
+
+  /**
+   * Triggers clean-up operations after a workspace is published.
+   *
+   * @param \Drupal\workspaces\Event\WorkspacePublishEvent $event
+   *   The workspace publish event.
+   */
+  public function onPostPublish(WorkspacePublishEvent $event): void {
+    $this->deleteAssociations($event->getWorkspace()->id());
   }
 
 }
